+++ packages/astro/src/index.ts
@@ -52,7 +52,7 @@ const createPlugin = (options?: SpotlightAstroIntegrationOptions): AstroIntegrat
           // will lead to both of them being enabled by the time this hook is called.
           // Setting one of them to `false` will not set the other one to false.
           // Therefore, both of them have to be `true` that we know that the toolbar is in fact active.
-          const hasToolbarEnabled = config.devToolbar?.enabled && config.devOverlay?.enabled;
+          const hasToolbarEnabled = config.devToolbar?.enabled || config.devOverlay?.enabled;
 
           // Before Astro 4, `devOverlay` was disabled by default and under `experimental`
           const hasExperimentalDevOverlayEnabled = !!(config as AstroConfigWithExperimentalDevOverlay).experimental
@@ -78,16 +78,18 @@ const createPlugin = (options?: SpotlightAstroIntegrationOptions): AstroIntegrat
         }
       },
 
-      'astro:server:start': async () => {
+      'astro:server:start': async ({ logger }) => {
         if (options?.sidecarUrl) {
+          logger.debug('Detected custom sidecar URL. Skipping default sidecar setup.');
           // If users set a custom sidecar URL, we assume they started the sidecar manually outside of Astro.
           // So we don't setup the default sidecar instance.
           return;
         }
+
         // Importing this dynamically because for some reason, the top level import
         // caused a dev server error because the sidecar code was bundled into the server
         const { setupSidecar } = await import('@spotlightjs/sidecar');
-        setupSidecar();
+        setupSidecar({ logger });
       },
     },
   };