+++ packages/overlay/src/constants.ts
@@ -1,0 +1,7 @@
+export const DEFAULT_SIDECAR_URL = 'http://localhost:8969/stream';
+
+export const DEFAULT_EXPERIMENTS = {
+  'sentry:focus-local-events': true,
+};
+
+export const DEFAULT_ANCHOR = 'bottomRight';