+++ packages/overlay/src/integrations/sentry/components/EventList.tsx
@@ -1,10 +1,15 @@
+import { useState } from 'react';
 import { Link } from 'react-router-dom';
-import CardList from '~/components/CardList';
+import { useSpotlightContext } from '~/lib/useSpotlightContext';
+import Badge from '../../../components/Badge';
+import CardList from '../../../components/CardList';
+import TimeSince from '../../../components/TimeSince';
 import { useSentryEvents } from '../data/useSentryEvents';
+import { useSentryHelpers } from '../data/useSentryHelpers';
 import { SentryEvent } from '../types';
 import { EventSummary } from './Events/Event';
+import HiddenItemsButton from './HiddenItemsButton';
 import PlatformIcon from './PlatformIcon';
-import TimeSince from './TimeSince';
 
 function renderEvent(event: SentryEvent) {
   return <EventSummary event={event} />;
@@ -12,12 +17,31 @@ function renderEvent(event: SentryEvent) {
 
 export default function EventList() {
   const events = useSentryEvents();
+  const helpers = useSentryHelpers();
+  const context = useSpotlightContext();
 
   const matchingEvents = events.filter(e => e.type !== 'transaction');
 
+  const [showAll, setShowAll] = useState(!context.experiments['sentry:focus-local-events']);
+  const filteredEvents = showAll
+    ? matchingEvents
+    : matchingEvents.filter(
+        e => (e.contexts?.trace?.trace_id ? helpers.isLocalToSession(e.contexts?.trace?.trace_id) : null) !== false,
+      );
+  const hiddenItemCount = matchingEvents.length - filteredEvents.length;
+
   return matchingEvents.length !== 0 ? (
     <CardList>
-      {matchingEvents.map(e => {
+      {hiddenItemCount > 0 && (
+        <HiddenItemsButton
+          itemCount={hiddenItemCount}
+          onClick={() => {
+            setShowAll(true);
+          }}
+        />
+      )}
+      {filteredEvents.map(e => {
+        const traceId = e.contexts?.trace?.trace_id;
         return (
           <Link
             className="hover:bg-primary-900 flex cursor-pointer items-center gap-x-4 px-6 py-2"
@@ -26,7 +50,13 @@ export default function EventList() {
           >
             <PlatformIcon event={e} className="text-primary-300" />
             <div className="text-primary-300 flex w-48 flex-col truncate font-mono text-sm">
-              <span>{(e.event_id || '').substring(0, 8)}</span>
+              <div className="flex items-center gap-x-2">
+                <div>{(e.event_id || '').substring(0, 8)}</div>
+                {traceId && helpers.isLocalToSession(traceId) ? (
+                  <Badge title="This event is part of your local session.">Local</Badge>
+                ) : null}
+              </div>
+              <span></span>
               <TimeSince date={e.timestamp} />
             </div>
             <div className="flex-1">{renderEvent(e)}</div>